Create Mangling & non public
MANGLING - using 2 underscores as for __balance in this case causes python to mangle the attribute and make it non-public to avoid clients from accidentally changing it. It is renamed to _Account__balance. Adding two underscores following balance as such __balance__ removes the mangling effects. Note that clients can still change it if they purposefully want to.