Made some slight modifications but heres the just if the programme -->Instances of song class are saved within album class as album.track, using add_song() method. Instances of Album object are saved in Artist.albums list using add_album() method, and instances of the artist class are saved in the list called artist list.
new_album, new_artist and new_song are used to save the new respective type of object in each new line of code. The fields are not saved in lists, the ENTIRE object is. This is really cool dip into OOP and also my first time using pycharm to access github directly. :D