OOP update
I have changed the programme such that it actually uses OOP. Encapsulation is shown here. The code is much simpler at the load_data function. The original code is so contrived.. After splitting each line in file to artist, album, year, song, we check if artist exists in the artist list. 
If not , we create artist object, and add object to artist list. 
We encapsulate all the adding of albums and songs within the Artist class, analogy being that if car is broken, you pass it to mechanic, not your issue anymore.
Artist class will handle, and with bit of fore sight we call the add_song method on artist before creating it.
Going to artist class, the add_song method directly adds songs to albums if they exist and creates them if they don't - Artist class knows about the existence of albums
but not songs. So it repeats the process of checking and creating if not found. It then calls on another add_song method for album class - this is further layer of encapsulation of song methods within album class.
This add song method repeats the process, but the parameters need to be referenced from album class encapsulating it.

Lastly, getters are introduced here. for song class, the song names are initiated as self.title. so  we created getter and used name = property(get_title) to add a property called name to our song class. We don't have to change any data attribute names but find_object method can work now (the method uses object.name for second parameter, but song has no such attribute). Will go through setters and getters in more detail