First time uploading a file instead
This makes use of both Composition and Inheritance.
basic classes all inherit from tag - I can say that head, body and DocType 'is a' tag. When an object is said to be a subset type of another object, should model using inheritance

HTML has a Doctype and has a body and has a head. When an object consist of a certain element/attribute of another, we should model using composition. Simply stated, HTML is composed of these 3 constituents