Create Parenthesis validity
Checks if parenthesis are in correct order of usage. Uses a dictionary with keys of left parenthesis and value of right parenthesis. Every time new right parenthesis is around, its added to the list. Every time right parenthesis is found, the last left parenthesis is popped of the list and value mapped to it is checked if it is equal the the right parenthesis found.
Theory behind it is that the sequence at which parenthesis are opened are also the sequenced in which they must be closed. Cannot have a single right parenthesis if a left one is still open. When a parenthesis is closed, pop it of the list.